notify-send: wait timeout variable should be int, not long
(cherry picked from commit 48f2d34c14aed208de6cf0ecae48ffeacfe40565)

+++ tools/notify-send.c
@@ -251,7 +251,7 @@ main (int argc, char *argv[])
         static gboolean     hint_error = FALSE, show_error = FALSE;
         static gboolean     transient = FALSE;
         static gboolean     wait = FALSE;
-        static glong        expire_timeout = NOTIFY_EXPIRES_DEFAULT;
+        static int          expire_timeout = NOTIFY_EXPIRES_DEFAULT;
         GOptionContext     *opt_ctx;
         GError             *error = NULL;
         gboolean            retval;
